Move cache and http core classes back to core
Core should contain the classes needed to write instrumentation, and binders should have instrumentation that we maintain. Therefore, the `CacheMeterBinder` and HTTP-related classes should remain in core. Related to fixing the cache stuff, this also moves the cache TCK implementations to micrometer-binders where the implementations of the cache binders are.
