+++ implementations/micrometer-registry-health/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
     api project(':micrometer-core')
-    api project(':micrometer-binders')
+    implementation project(':micrometer-binders')
 
     implementation 'org.slf4j:slf4j-api'
