+++ implementations/micrometer-registry-statsd/build.gradle
@@ -4,7 +4,7 @@ plugins {
 
 dependencies {
     api project(':micrometer-core')
-    api project(':micrometer-binders')
+    implementation project(':micrometer-binders')
 
     implementation 'io.projectreactor:reactor-core'
     implementation('io.projectreactor.netty:reactor-netty-core') {
