+++ micrometer-binders/build.gradle
@@ -1,14 +1,6 @@
-plugins {
-    id 'idea'
-}
+description 'Instrumentation of components using Micrometer'
 
 dependencies {
-    // TODO(anuraaga): HdrHistogram is exposed in the micrometer API but probably shouldn't be
-    api 'org.hdrhistogram:HdrHistogram'
-    implementation('org.latencyutils:LatencyUtils') {
-        exclude group: 'org.hdrhistogram', module: 'HdrHistogram'
-    }
-
     api project(':micrometer-core')
 
     // cache monitoring
@@ -40,7 +32,6 @@ dependencies {
     optionalApi 'ch.qos.logback:logback-classic'
     optionalApi 'org.apache.logging.log4j:log4j-core'
 
-    // @Timed AOP
     optionalApi 'com.squareup.okhttp3:okhttp'
 
     optionalApi 'org.mongodb:mongodb-driver-sync'
@@ -49,6 +40,8 @@ dependencies {
 
     optionalApi 'org.apache.kafka:kafka-clients'
     optionalApi 'org.apache.kafka:kafka-streams'
+
+    testImplementation project(':micrometer-test')
 
     testImplementation 'io.projectreactor:reactor-test'
 
@@ -102,6 +95,8 @@ dependencies {
     testImplementation 'org.postgresql:postgresql'
 
     testImplementation 'org.testcontainers:mongodb'
+
+    testImplementation 'org.jsr107.ri:cache-ri-impl'
 }
 
 task shenandoahTest(type: Test) {
