+++ micrometer-binders/src/main/java/io/micrometer/binder/cache/HazelcastCacheMetrics.java
@@ -27,6 +27,7 @@ import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.Tags;
 import io.micrometer.core.instrument.binder.BaseUnits;
 import io.micrometer.binder.cache.HazelcastIMapAdapter.LocalMapStats;
+import io.micrometer.core.instrument.binder.cache.CacheMeterBinder;
 import io.micrometer.core.lang.NonNullApi;
 import io.micrometer.core.lang.NonNullFields;
 import io.micrometer.core.lang.Nullable;
