+++ micrometer-binders/src/test/java/io/micrometer/binder/cache/CaffeineCacheMetricsCompatibilityTest.java
@@ -13,49 +13,44 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.core.instrument.binder.cache;
+package io.micrometer.binder.cache;
 
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
+import com.github.benmanes.caffeine.cache.Caffeine;
+import com.github.benmanes.caffeine.cache.LoadingCache;
+import io.micrometer.core.instrument.binder.cache.CacheMeterBinder;
+import io.micrometer.core.instrument.binder.cache.CacheMeterBinderCompatibilityKit;
 
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.atomic.AtomicReference;
 
 import static java.util.Collections.emptyList;
 
-class GuavaCacheMetricsCompatibilityKit extends CacheMeterBinderCompatibilityKit<LoadingCache<String, String>> {
+class CaffeineCacheMetricsCompatibilityTest extends CacheMeterBinderCompatibilityKit<LoadingCache<String, String>> {
     private AtomicReference<String> loadValue = new AtomicReference<>();
 
     @Override
     public LoadingCache<String, String> createCache() {
-        return CacheBuilder.newBuilder()
+        return Caffeine.newBuilder()
                 .maximumSize(2)
                 .recordStats()
-                .build(new CacheLoader<String, String>() {
-                    @Override
-                    public String load(String key) throws Exception {
-                        String val = loadValue.getAndSet(null);
-                        if (val == null)
-                            throw new Exception("don't load this key");
-                        return val;
-                    }
+                .executor(Runnable::run)
+                .build(key -> {
+                    String val = loadValue.getAndSet(null);
+                    if (val == null)
+                        throw new Exception("don't load this key");
+                    return val;
                 });
     }
 
     @Override
     public CacheMeterBinder<LoadingCache<String, String>> binder() {
-        return new GuavaCacheMetrics<>(cache, "mycache", emptyList());
+        return new CaffeineCacheMetrics<>(cache, "mycache", emptyList());
     }
 
     @Override
     public void put(String key, String value) {
         synchronized (this) {
             loadValue.set(value);
-            try {
-                cache.get(key);
-            } catch (ExecutionException ignored) {
-            }
+            cache.get(key);
         }
     }
