+++ micrometer-binders/src/test/java/io/micrometer/binder/cache/EhCache2MetricsTest.java
@@ -21,7 +21,7 @@ import io.micrometer.core.instrument.FunctionCounter;
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Tags;
-import io.micrometer.binder.cache.EhCache2Metrics;
+import io.micrometer.core.instrument.binder.cache.AbstractCacheMetricsTest;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import net.sf.ehcache.Cache;
 import net.sf.ehcache.CacheManager;
