+++ micrometer-binders/src/test/java/io/micrometer/binder/cache/JCacheMetricsCompatibilityTest.java
@@ -13,8 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.core.instrument.binder.cache;
+package io.micrometer.binder.cache;
 
+import io.micrometer.core.instrument.binder.cache.CacheMeterBinder;
+import io.micrometer.core.instrument.binder.cache.CacheMeterBinderCompatibilityKit;
 import org.jsr107.ri.spi.RICachingProvider;
 
 import javax.cache.Cache;
