+++ micrometer-binders/src/test/java/io/micrometer/binder/jetty/TimedHandlerTest.java
@@ -31,8 +31,7 @@ import javax.servlet.http.HttpServletResponse;
 
 import io.micrometer.core.instrument.MockClock;
 import io.micrometer.core.instrument.Tags;
-import io.micrometer.binder.http.Outcome;
-import io.micrometer.binder.jetty.TimedHandler;
+import io.micrometer.core.instrument.binder.http.Outcome;
 import io.micrometer.core.instrument.simple.SimpleConfig;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import org.eclipse.jetty.http.HttpStatus;
