+++ micrometer-core/src/main/java/io/micrometer/core/instrument/binder/cache/package-info.java
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 @NonNullApi
-@Deprecated
 package io.micrometer.core.instrument.binder.cache;
 
 import io.micrometer.core.lang.NonNullApi;
