+++ micrometer-core/src/main/java/io/micrometer/core/instrument/binder/http/DefaultHttpServletRequestTagsProvider.java
@@ -25,12 +25,10 @@ import javax.servlet.http.HttpServletResponse;
 /**
  * Default {@link HttpServletRequestTagsProvider}.
  *
- * @deprecated Scheduled for removal in 2.0.0, please use {@code io.micrometer.binder.http.DefaultHttpServletRequestTagsProvider}
  * @author Jon Schneider
  * @since 1.4.0
  */
 @Incubating(since = "1.4.0")
-@Deprecated
 public class DefaultHttpServletRequestTagsProvider implements HttpServletRequestTagsProvider {
     @Override
     public Iterable<Tag> getTags(HttpServletRequest request, HttpServletResponse response) {
