+++ micrometer-core/src/main/java/io/micrometer/core/instrument/binder/http/HttpRequestTags.java
@@ -25,12 +25,10 @@ import javax.servlet.http.HttpServletResponse;
 /**
  * Tags for HTTP requests.
  *
- * @deprecated Scheduled for removal in 2.0.0, please use {@code io.micrometer.binder.http.HttpRequestTags}
  * @author Jon Schneider
  * @since 1.4.0
  */
 @Incubating(since = "1.4.0")
-@Deprecated
 public class HttpRequestTags {
     private static final Tag EXCEPTION_NONE = Tag.of("exception", "None");
