+++ micrometer-core/src/main/java/io/micrometer/core/instrument/binder/http/HttpServletRequestTagsProvider.java
@@ -24,13 +24,11 @@ import javax.servlet.http.HttpServletResponse;
 /**
  * Provides {@link Tag Tags} for HTTP Servlet request handling.
  *
- * @deprecated Scheduled for removal in 2.0.0, please use {@code io.micrometer.binder.http.HttpServletRequestTagsProvider}
  * @author Jon Schneider
  * @since 1.4.0
  */
 @Incubating(since = "1.4.0")
 @FunctionalInterface
-@Deprecated
 public interface HttpServletRequestTagsProvider {
     /**
      * Provides tags to be associated with metrics for the given {@code request} and
