+++ micrometer-core/src/main/java/io/micrometer/core/instrument/binder/http/package-info.java
@@ -13,5 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@Deprecated
 package io.micrometer.core.instrument.binder.http;
