+++ micrometer-test/src/main/java/io/micrometer/core/instrument/binder/cache/AbstractCacheMetricsTest.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.binder.cache;
+package io.micrometer.core.instrument.binder.cache;
 
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.Tags;
-import io.micrometer.binder.cache.CacheMeterBinder;
+import io.micrometer.core.instrument.binder.cache.CacheMeterBinder;
 import io.micrometer.core.instrument.search.RequiredSearch;
 
 /**
@@ -26,7 +26,7 @@ import io.micrometer.core.instrument.search.RequiredSearch;
  *
  * @author Oleksii Bondar
  */
-abstract class AbstractCacheMetricsTest {
+public abstract class AbstractCacheMetricsTest {
 
     protected Tags expectedTag = Tags.of("app", "test");
