+++ micrometer-test/src/main/java/io/micrometer/core/instrument/binder/cache/CacheMeterBinderCompatibilityKit.java
@@ -26,7 +26,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 /**
  * @author Jon Schneider
  */
-@Deprecated
 public abstract class CacheMeterBinderCompatibilityKit<C> {
     private MeterRegistry registry = new SimpleMeterRegistry();
     private CacheMeterBinder<C> binder;
