+++ samples/micrometer-samples-javalin/src/main/java/io/micrometer/javalin/samples/PrometheusSample.java
@@ -23,7 +23,6 @@ import io.javalin.http.HandlerType;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.Tags;
-import io.micrometer.binder.http.DefaultHttpServletRequestTagsProvider;
 import io.micrometer.binder.jetty.JettyConnectionMetrics;
 import io.micrometer.binder.jetty.JettyServerThreadPoolMetrics;
 import io.micrometer.binder.jetty.TimedHandler;
@@ -32,6 +31,7 @@ import io.micrometer.binder.jvm.JvmHeapPressureMetrics;
 import io.micrometer.binder.jvm.JvmMemoryMetrics;
 import io.micrometer.binder.system.FileDescriptorMetrics;
 import io.micrometer.binder.system.ProcessorMetrics;
+import io.micrometer.core.instrument.binder.http.DefaultHttpServletRequestTagsProvider;
 import io.micrometer.core.instrument.util.StringUtils;
 import io.micrometer.core.lang.NonNull;
 import io.micrometer.prometheus.PrometheusConfig;
